// src/types/Request.ts
export type Request = {
  id: string
  assignee_avatar: string
  assignee_id: string
  assignee_name: string
  comment_count: number
  created_date: any
  description: string
  dislike_count: number
  gps_location: any
  image: string
  last_responded: string
  likes_count: string
  priority: string
  resident_avatar: string
  resident_id: string
  resident_name: string
  status: string
  sub_type: string
  sub_type_icon: string
  type: string
  type_icon: string
  formatted_address: string
}